+++ src/main/ml-modules/root/ds/lux/searchWillMatch.mjs
@@ -2,12 +2,11 @@ import { getObjectFromNode } from '../../utils/utils.mjs';
 import { handleRequest } from '../../lib/securityLib.mjs';
 import { determineIfSearchWillMatch } from '../../lib/searchLib.mjs';
 
-const unitName = external.unitName;
 const multipleSearchCriteria = getObjectFromNode(external.q);
 
 const response = handleRequest(function () {
   return determineIfSearchWillMatch(multipleSearchCriteria);
-}, unitName);
+});
 
 response;
 export default response;