+++ src/main/ml-modules/root/ds/lux/stats.mjs
@@ -1,8 +1,6 @@
 import { handleRequest } from '../../lib/securityLib.mjs';
 import { getScopeEstimates } from '../../lib/environmentLib.mjs';
 
-const unitName = external.unitName;
-
 const response = handleRequest(function () {
   const start = new Date();
   const end = new Date();
@@ -15,7 +13,7 @@ const response = handleRequest(function () {
       milliseconds: end - start,
     },
   };
-}, unitName);
+});
 
 response;
 export default response;